Update project3.s
Here we first load the base address of the user's input into $t1. Then reserve 4 bytes of storage in the run time stack( for a later usage). We then push into the stack the content of $t1 into the first byte address of the stack( which is the top in the running stack). We finally jump and link to "sub_a" sub  program.