Update project3.s
After the increment, we jump to "LoopSub".
Now we are under the Send_to_sub_b subprogram. We first save the word in $t1 into the 4th byte reading from the top to the bottom of the running stack. Then you jump and link to sub_b( which follows the requires as given in the instruction prompts).
Once its has jump to the return address and process to the next instruction, we load the word on top of the running stack into $t8. Then after, We branch to print the error message, "X" if the value in $t8 equals -1, a way to catch invalid input.