Update project3.s
Next jump to print_val. The next lines of instructions is to print the error message, "X" if there is an invalid input. First set up to print a character. Then we load "X" into $a0 after which we make a system call.
The final instruction is to jump to checking_pos sub program, which we will use to check current positions.