Update project3.s
Under the  print_val subprogram, include instructions to print a an integer onto the console. We fetch the value from $t8 which should be holding our expected integer into $a0 and then make a system call to execute the print command. Then after  we set up to print a string opening parenthesis.