Update project3.s
On the top, we make a system call to print out the opening parenthesis. Next we set up to print an integer which should be the number of characters based <= 4 and then we make a system call to execute.
The last line of instruction is to set up to print a character, this time the closing parenthesis.