Update project3.s
The first two lines of instructions completes pseudo instruction to print the closing parenthesis onto the console.  Next we jump to checking_pos subprogram.
Under the subprogram, we first load word from the 4th byte position of the run time stack from the top to the bottom into $t9. Next, we free up 8 bytes of storage which was formally reserved.