Update project3.s
Next, we load byte from $t9 into $t7. After, we branch to end subprogram, if $t7's content equals 10 or 0. Else if its value equals 44, we branch to next_substr subprogram.
The last two lines of instructions( and labels) is to get the ready for processing the next substring with the help of the comma's in between.