Update project3.s
Moving forward, we first increment the value in $t1 by 1(which serves as our counter). Then after, we set up to print the comma character for separating the corresponding answers for substring, the comma right after printing out the base N decimal of the current substring.
We finally invoke the system to print out onto the console and then after jump to Loop subprogram.