Update project3.s
In these lines of instructions, first we load on top of the stack( where the stack pointer is currently pointing) into $t2. Next we load the word on the 4th address on the run time stack reading from top to bottom into $t0 and finally we free up the 8 bytes that we reserve formerly.
Moving forward, we enter into the LoopLeading subprogram which allows us to loop the leading tab characters and continue ultimate with our looping.
To begin with the instruction, first loading the byte from $t2 into $t4.