Update project3.s
First we branch to For_Valid subprogram if the value in $t4 equals to that in $t0. Else if the value in $t4 equals 32, we branch to TabL subprogram. Otherwise we branch to SkipL.
The TabL subprogram branch to Loop_Main if the value in $t4 does not equal to 9 and the SkipL subprogram increase the value in $t2 by 1.