Update project3.s
Moving forward, we jump to LoopLeading, executing all lines of instructions under that label.
The next sub program is the Loop_Main subprogram .
First, we load the byte from $t2 into $t4.  Then we branch to For_Invalid subprogram if value in $t5 is greater than 4. Else if the value in $t4 equals that in $t0, we branch to For_Valid. Else if the value in $t4 equals 32, we branch to Loop_Trailing subprogram.