Update project3.s
Continuing, we branch to Loop_Trailing if the value in $t4 equals 9.
The number subprogram follows  which allows us to process the decimal equivalent for integer string between 0 and 9.
First, we check if the value in $t4 is greater than 57, if yes, then we branch to upper_case( since its a potential upper_case letter using ascii value). However if it's less than 48, then it surely must be invalid so we branch to For_Invalid. Otherwise, we $t3 to help us calculate for the decimal equivalence of the number string, hence jump to calculations sub program.