Update project3.s
The next sub program is upper_case. First we check if value in $t4 is greater than 86, if yes, we branch to lower_case subprogram( since there could be the next possibity for such an ascii value). Else if its less than 65, then it's surely invalid so we branch to For_Invalidl Otherwise we prepare the $t3 register for calculations.
The last instruction starts the instructions under the lower_case sub program.
If value in $t4 is greater than 118 considering my beta value, then its surely invalid so branch to For_Invalid.