Update project3.s
Continuing, if the value in $t4 is less than 97 f, it must be invalid so we branch to For_Invalid. Otherwise prepare a value in $t3 for calculation.
The following three instructions is mean to get the convert the character into their respective decimal value, add the value to $t8 for every iteration and then multiply by 32 for each iteration.