Update project3.s
Next is the move_to_next sub program  to advance to the next byte. We increase the value in $t5 by one, a counter and then loop again.
Underneath is also, the Loop_Trailing subprogram where we load the subsequent byte into $t4 and then branch to For_valid if val in $t4 equals value in $t0.