Update project3.s
Continuing we check if we have found trailing space and then we continue to loop by jumping to SkipT.
The TabT subprogram is there so check if a trailing tab was found. The SkipT subprogram allows us to go to the next byte address and finally we jump to Loop_Trailing.