Update project3.s
Under the Division subprogram, we divide $t8 by my N, 32. Next we reserve 8 bytes of storage in the run time stack. Then we push the value in $t8 into the top of the stack and then the value in $t9 into the 4th address location 4 bytes away from the top of the run time stack. Finally we jump to the return address.